fix: move SearchBox into topbar components
