+++ app/cli/npm/index.js
@@ -1,4 +1,3 @@
-import audit from './audit';
 import doctor from './doctor';
 import prune from './tooling/prune';
 import dedupe from './tooling/dedupe';
@@ -8,7 +7,6 @@ import init from './tooling/init';
 
 export default {
   init,
-  audit,
   doctor,
   prune,
   dedupe,
